+++ service/submitqueue/demo/pr/BUILD.bazel
@@ -2,17 +2,16 @@ load("@rules_go//go:def.bzl", "go_binary", "go_library", "go_test")
 
 go_library(
     name = "go_default_library",
-    srcs = ["main.go"],
+    srcs = [
+        "github.go",
+        "main.go",
+    ],
     importpath = "github.com/uber/submitqueue/service/submitqueue/demo/pr",
     visibility = ["//visibility:private"],
     deps = [
-        "//api/base/change/protopb:go_default_library",
         "//api/base/mergestrategy/protopb:go_default_library",
-        "//api/submitqueue/gateway/protopb:go_default_library",
         "//platform/base/change/github:go_default_library",
-        "//submitqueue/entity:go_default_library",
-        "@org_golang_google_grpc//:go_default_library",
-        "@org_golang_google_grpc//credentials/insecure:go_default_library",
+        "//submitqueue/client:go_default_library",
     ],
 )
 
@@ -27,9 +26,7 @@ go_test(
     srcs = ["main_test.go"],
     embed = [":go_default_library"],
     deps = [
-        "//api/submitqueue/gateway/protopb:go_default_library",
         "@com_github_stretchr_testify//assert:go_default_library",
         "@com_github_stretchr_testify//require:go_default_library",
-        "@org_golang_google_grpc//:go_default_library",
     ],
 )
