+++ service/submitqueue/gateway/client/BUILD.bazel
@@ -6,12 +6,8 @@ go_library(
     importpath = "github.com/uber/submitqueue/service/submitqueue/gateway/client",
     visibility = ["//visibility:private"],
     deps = [
-        "//api/base/change/protopb:go_default_library",
-        "//api/base/mergestrategy/protopb:go_default_library",
-        "//api/submitqueue/gateway/protopb:go_default_library",
         "//platform/base/change/github:go_default_library",
-        "@org_golang_google_grpc//:go_default_library",
-        "@org_golang_google_grpc//credentials/insecure:go_default_library",
+        "//submitqueue/client:go_default_library",
     ],
 )
